requirement.txt
python-dotenv
streamlit
langchain-groq
langchain 
Groq 
langchain_core